refactor(mentor): 파일명 .js -> .jsx 수정
