+++ components/BoardDetailComponents/BoardDetailInfo.jsx
