+++ components/BoardDetailComponents/ChatItem.jsx
