+++ components/ItemComponents/ItemList.jsx
@@ -2,8 +2,8 @@ import React from "react";
 import { formatPrice } from "@/utils/price";
 import styles from "./ItemList.module.css";
 import img_default from "@/images/img_default.png";
-import { MobileSearchBar } from "./MobileSearchBar";
-import { DesktopSearchBar } from "./DesktopSearchBar";
+import { MobileSearchBar } from "./MobileSearchBar.jsx";
+import { DesktopSearchBar } from "./DesktopSearchBar.jsx";
 import Link from "next/link";
 import { ROUTES } from "@/utils/rotues";
 import { useDeviceType } from "@/hooks/useDeviceType";
