+++ components/ItemComponents/ItemListContainer.jsx
@@ -1,6 +1,6 @@
 import React, { useState } from "react";
 import { useProducts } from "@/hooks/useProducts";
-import ItemList from "./ItemList";
+import ItemList from "./ItemList.jsx";
 
 export default function ItemListContainer() {
   const [keyword, setKeyword] = useState("");
