+++ pages/createBoard.jsx
@@ -1,4 +1,4 @@
-import CreateForm from "@/components/CreateBoardComponents/CreateForm";
+import CreateForm from "@/components/CreateBoardComponents/CreateForm.jsx";
 export default function Board() {
   return (
     <>
