+++ pages/createProduct.jsx
@@ -1,4 +1,4 @@
-import CreateBtn from "@/components/CreateProductComponents/CreateBtn";
+import CreateBtn from "@/components/CreateProductComponents/CreateBtn.jsx";
 import styles from "@/styles/createProduct.module.css";
 
 export default function createProduct() {
