+++ pages/signin.jsx
@@ -1,4 +1,4 @@
-import SigninForm from "../components/SigninComponents/SigninForm";
+import SigninForm from "../components/SigninComponents/SigninForm.jsx";
 import styles from "../styles/signin.module.css";
 
 export default function Signin() {
